# Conflicts:
#	.idea/dataSources.local.xml
#	.idea/dataSources/1b68a51e-6688-4d38-b10c-b4aaec735fd3.xml
#	.idea/dataSources/1b68a51e-6688-4d38-b10c-b4aaec735fd3/storage_v2/_src_/database/syllabus.5Q1kaQ/schema/public.abK9xQ.meta
#	.idea/workspace.xml
#	src/main/java/kz/syllabus/repository/InstructorRepository.java

+++ src/main/java/kz/syllabus/repository/DisciplineInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DisciplineInfoRepository extends JpaRepository<DisciplineInfo, Long> {
+public interface DisciplineInfoRepository extends JpaRepository<DisciplineInfo, Integer> {
 
 }
